Refactor of kernels

+++ BoidFlockSimulation/WindowSDL.cpp
@@ -66,7 +66,7 @@ int WindowSDL::addBoidToRenderer(float4 boid, int boidSize)
 	float2 position = Calculator::getBoidPosition(boid);
 	float angle = Calculator::getAngleFromVector(Calculator::getBoidVelocity(boid));
 
-	SDL_Rect dstrect = { position.x, position.y, boidSize, boidSize };
+	SDL_Rect dstrect = { (int)position.x, (int)position.y, boidSize, boidSize };
 	SDL_RendererFlip flip = (SDL_RendererFlip)(SDL_FLIP_NONE);
 
 	if (SDL_RenderCopyEx(_renderer, _boidTexture, NULL, &dstrect, angle, NULL, flip) < 0)
